Added the object oriented planning sheet to the repo
As previously mentioned, milestone 4 is there as a bumper in the case things go very, very wrong. Needless to say, things haven't gone quite as well as I had planned as you might be able to see from the lack of commits as of last week. I plan to finish a few more things before class today of course, but there just hasn't been as much time as I had hoped given the amount of assignments in other classes.
Hopefully its not too much of an issue though. everything else should be going as planned.