The program now runs!
Fixed a bunch of bugs that were preventing the program from running
So far the only thing left to fix would be the super enemy list, which for some reason, has yet to work properly.
That will be my next commit as soon as I am able to do it.