Finished Milestone 1
Yes, this is milestone 1. It should include a grand total of 26 different inventory skills, due to the large amount of changes the inner workings of the code underwent.
Next, I plan to add fleshy, and hopefully add fleshy as well as the other enemies. These will come in commits as soon as I have them done.